+++ app/src/main/java/app/netlify/dev_ali_hassan/hafizalquran/api/QuranApiHelper.kt
@@ -4,5 +4,5 @@ import app.netlify.dev_ali_hassan.hafizalquran.api.dataclasses.QuranApiResponse
 import retrofit2.Response
 
 interface QuranApiHelper {
-    suspend fun getPageByNumber(pageNumber: Long): Response<QuranApiResponse>
+    suspend fun getPageByNumber(pageNumber: Int): Response<QuranApiResponse>
 }
